+++ app/models/engine/interpreter.rb
@@ -115,13 +115,14 @@ module Engine
       hash = {}
       hash[:variable] = variable
       hash[:sheets] = { subject_event: SubjectEvent.where(event: event) } if event
-      svs = SheetVariable
+      klass = (variable.variable_type == "checkbox" ? Response : SheetVariable)
+      values = klass
         .left_outer_joins(:domain_option).joins(:sheet).merge(Sheet.current).where(hash).order(:sheet_id)
-        .pluck(:subject_id, :sheet_id, :design_id, domain_option_value_or_value, :missing_code)
+        .pluck(:subject_id, :sheet_id, :design_id, domain_option_value_or_value(table: klass.table_name), :missing_code)
       formatter = Formatters.for(variable)
       integer_regex = Regexp.new(/^[-+]?[1-9][0-9]*$/)
       decimal_regex = Regexp.new(/^[-+]?[0-9]*(\.)?[0-9]+$/)
-      svs.each do |subject_id, sheet_id, design_id, value, missing_code|
+      values.each do |subject_id, sheet_id, design_id, value, missing_code|
         formatted_value = formatter.raw_response(value)
         if formatted_value.is_a?(String) && (formatted_value == "0" || !(integer_regex =~ formatted_value).nil?)
           formatted_value = Integer(formatted_value)